refactor: extract gateway admin projections
