+++ crates/dcc-mcp-gateway-admin/src/lib.rs
@@ -1,13 +1,15 @@
 //! Domain and embedded frontend boundary for the DCC-MCP gateway admin dashboard.
 //!
-//! This crate owns admin-facing trace and caller-context contracts independently
-//! of gateway routing state. It also owns the Vite/npm build script and generated
-//! dashboard payload; the Node.js toolchain only runs when `embed` is enabled.
+//! This crate owns admin-facing trace, caller-context, and compact projection
+//! contracts independently of gateway routing state. It also owns the Vite/npm
+//! build script and generated dashboard payload; the Node.js toolchain only runs
+//! when `embed` is enabled.
 
 #![forbid(unsafe_code)]
 
 /// Admin trace and caller-attribution value types.
 pub mod domain;
+mod projection;
 mod trace_log;
 
 pub use domain::agent_context::{
@@ -20,6 +22,10 @@ pub use domain::trace::{
     MAX_AGENT_CONTEXT_STRING_BYTES, MAX_INPUT_BYTES, MAX_OUTPUT_BYTES, TOKEN_ESTIMATOR,
     TokenTelemetry, TraceContext, TraceContextHeader, TracePayload, TraceSpan, estimate_tokens,
     parse_traceparent,
+};
+pub use projection::{
+    compact_debug_bundle_payload, compact_trace_context_payload, compact_trace_detail_payload,
+    compact_trace_list_payload,
 };
 pub use trace_log::TraceLog;
